Update master branch to guidebot v13 (#144)
* Initial update to v13

Updated everything

* Update guildCreate.js

updated from `user.id` to `ownerId`

* V13 Setup.JS fix + reboot.JS command update (#139)

* Adding partials to update baseConfig

* Adding logical "restart" alias

* Updated commands.

Updated all the commands that output codeblocks, used the new formatters.

* Thread support

Added a single line of code to make Guidebot auto-join any threads created, useful for moderation bots.

* Guidebot will now ignore disabled commands

Fix #137

* Updating invite link

* Removed unused dep

* Slash Commands (#143)

* Slash Commands

Added slash command support to the original (main) version of guidebot.

Updated discord.js to version 13.0.1, updated the throw in the index to require version 16 of nodejs.

Brought both versions of guidebot to the same functional level.

* Fixed a small mistake

The setup placed the partial array inside an array causing an error.

* Minor update

Updated the readme to be concise with the other version.
Updated the setup script, removed the deprecated ownerId field.

* Consistency fix

The prefix for the bot in two files were different compared to the one in the set up script. Fixed for consistency.

* The real quick fix

Actually fixed the damn thing this time.